Sign up form
Right now we are not able to provide users default signup form. We already have one form in mail module but it is more focused to email subscriptions.

Added possibility to register user to crm with possibility to redirect to referal page.

Main motivation is to provide singup form for sales funnel that require login (tyzden#259)