Making sales funnel final routes cover error cases
When eshop payment didn't end up with success, eshop router wasn't
able to find the correct route to sales funnel module and executed
the action with default shop route.

    new Route('//' . $shopHost . '/<action>[/<id>[/<code>]]', 'Products:Shop:default');

This ended up with 404 as sales funnel error could not be handled with
this route. By explicitly adding route to SalesFunelModule, all
payment result cases are covered for eshop.

remp/crm#775