Changed behavior of admin locale switcher
The switcher no longer stores language information to session, but sets
the locale directly to the user account. Available locales are now
initialized when the component is rendered, opposed to the constructor
in the previous implementation.

Composer scripts were initializing widgets as well due to the missing
$_SERVER['SHELL'] variable. A fix using $_SERVER['SHLVL'] variable
was put in place and the flow was altered a bit to avoid unnecessary
widget initialization.

remp/crm#2240