Add label for default sales funnel to list of funnels
Default sales funnel is set in CRM admin config and there was no
indication of it in list of funnels or detail of funnel. To check which
is default (visible at https://crm.press/subscriptions), admin had to go
into CRM admin configuration page.

This is just visual change.

remp/crm#2830