Added authorized status to list of included statuses for conversions
If funnel used preauthorization gateway for one of its trials,
the funnels' "conversion" metric was not being incremented
for these kind of payments, eventhough the process was successful,
user is a subscriber and will be charged during renewal.

This wasn't correct and preauthorization within the funnel
should be counted as a funnel's conversion.

remp/novydenik#1413